requirements.txt

--index-url https://pypi.org/simple
pyqt6==6.8.1                                 # can be installed in general system
--extra-index-url https://download.pytorch.org/whl/rocm6.2.4
torch==2.6.0+rocm6.2.4                 # can be installed in general system
python-binance # used ==1.0.28